Implement IPv6 in symbian QNetworkInterface
For IPv6 interfaces, the network mask is reported correctly, so we use
that.
For IPv4 interfaces, the network mask is reported as 0.0.0.0 so the
existing workaround to get the netmask from routes is used.
For loopback interfaces, the mask can be statically determined.

Skip checking the routes if proper masks were reported for all the
interfaces.

Added a helper function for converting TInetAddr -> QHostAddress using
the constructors that take binary data. (better than conversion to/from
strings)

Task-number: QTBUG-18137
Reviewed-by: Markus Goetz