Add flag for forcibly propagating backing store alpha to framebuffer
Previously, the following rules applies to the creation and blitting of
the Symbian raster backing store:

    1. Creation of backing store with an alpha channel:
	   Backing store has an alpha channel only if !QWidget::isOpaque.

	2. Pre-filling of backing store prior to paint loop:
	   Backing store is filled with transparent pixels if
	   !QWidget::isOpaque.

	3. Blitting of backing store:
	   CGraphicsContext::EDrawModeWriteAlpha is used (meaning that backing
	   store alpha values are propagated into the frame buffer), if
	   QWidget::isOpaque.

In order for the direct camera viewfinder to be visible on DSA devices,
alpha=0 must be written into the framebuffer in the region where the
viewfinder will be displayed.  This requires a backing store with an alpha
channel (1), use of CGraphicsContext::EDrawModeWriteAlpha (3), but not
pre-filling of the entire backing store (2).

This patch adds a new enum value, QWExtra::BlitWriteAlpha,  which can be
used by camera backends to achieve the desired behaviour.

Task-number: QTMOBILITY-1278
Reviewed-by: Jani Hautakangas