Make tst_QEventLoop::processEventsExcludeSocket stricter
The test was only checking that the socket event is not processed when
excluded. I have added an additional test that a further call with socket
events included results in the deferred socket event being processed.
i.e. makes sure events are deferred to later and not lost.

Also fixed capabilities of the autotest, so this test case can run.

Reviewed-by: Markus Goetz