Allow import functions to report multiple errors
Previously, the functions in QDeclarativeImport (and related classes)
took an errorString ptr argument, which they could fill with an error.
However, in some cases it is better if the function can provide more
than one error, and more information (eg, parsing qmldir files can
result in multiple errors, with specific line/column information).

This commit changes those functions so that they take a pointer to a
list of QDeclarativeError as an argument, so that the functions can
provide the caller with accurate error information.

It also reports errors which occur during parsing of qmldir or plugin
loading caused by implicit import of ".".

Task-number: QTBUG-17530
Task-number: QTBUG-16894
Reviewed-by: Aaron Kennedy
Change-Id: I783191a99aeebba9f323e508cf92b45017c6aa4b