Fix bug in scarce resource handling
In one particular codepath (where a JS function in a QML file is
invoked directly from CPP) the scarce resources were not referenced
and dereferenced correctly (in QDeclarativeVMEMetaObject).

This commit adds referencing and dereferencing to that code path,
and adds a unit test to ensure that we do not regress.

Relates to commit: 59ace5c5a666b1588560d2aeaa79a57da535e863
Reviewed-by: Aaron Kennedy
Task-number: QMLNG-18
Change-Id: Ic5f59ec4d371b53a705f5f10291d49af9eb0c3fd