handle errors in QSymbianHostResolver::requestHostLookup
It was hanging because although the error flags were set, the callback
was not emitted and nothing was done with the return value from the
function.
Now, it sets the state to error, and self completes - so that the
completion through RunL is used as normal. Because the state is error,
the processing of name/address is skipped and the error string is not
rewritten.
Note, error being detected at this point is only common when using an
explicit network session (as RHostResolver::Open can fail if the
RConnection is in the wrong state)

Reviewed-by: Markus Goetz