qmake vcproj generator: do not insert $(INHERIT)
This will remove the warnings about undefined environment variables
when building with IncrediBuild.
The $(INHERIT) feature does not work in the "Create PCH through file"
property of Visual Studio. It expands to an empty string but doesn't
yield a warning in a standard VS build.
Also, the value of "Create PCH through file" is supposed to be exactly
the string as in the include statement of the cpp file.

Done-with: Marius Storm-Olsen