Add a feature to QTestLib to correct benchmark results.
Two special values are added to the measurement of benchmarks: a zero
and a comparison. The "zero" value is to be used when the test incurs
a non-negligible overhead just to run. The benchmark author should add
an "empty" run that measures basically the overhead and set it to
attribute QTest::Zero.

The "zero" value will be subtracted from any benchmark measurements
before reporting (currently the plain logger only). Note that all
benchmark types don't make sense, notably those already measurement
some kind of rate, like frames per second and bytes per second.

The "baseline" value is the comparison (with the "zero"
subtracted"). When the baseline is set, the library will instead
report the ratio between the benchmark and the baseline, in percentage
or a multiple (e.g. "75%" [of the time] or "10.0x" [faster]).

Adding measurements in "robes" left for future improvement.

This feature is not reviewed and will be reverted later.