Fix a deadlock in symbian QHostInfo when aborting
DoCancel was calling lookupFinished, which tries to acquire a mutex again
that was already acquired in abortLookup (causing a deadlock).
This call is un-necessary, as Cancel was only called from the destructor
and from abortLookup - in neither case is this wanted.

The run and RunError functions explicitly call lookupFinished.

This fixes hang in autotests when attempting to lookup a bad host name.

Reviewed-by: Markus Goetz