Designer: Avoid nested QDialog's in item widget content dialogs.
Refactor the AbstractItemEditor to be a QWidget instead of
a dialog as it is used for the Rows/Columns of trees, tables as
well, creating a hierarchy of nested QDialogs. The nested
QDialogs cause some of the crashes observed with transparent
styles using event filters accessing QWidget::window() in QEvent::Hide
event filters.

Task-number: QTBUG-18222