Implement tiled image and pixmap drawing in VG paint engine.
The vgWritePixel-based "fallback" in drawImage can only be used when
the transformation is simple and no opacity is set. The tiled drawing
introduced here will serve as an additional fallback for the more
complex cases and will allow drawing large (12 Mpix or more) pixmaps
and images with transformation and semi-transparency even when the
complete image data would not fit into the GPU memory.

Task-number: QTBUG-18251
Reviewed-by: Jani Hautakangas