Add ARM Neon versions of fromLatin1 and fromUtf8
The fromLatin1 code is very simple, yet the handwritten assembly
performs better due to the use of post-increments.

The fromUtf8 code has two alternatives. Neon lacks an instruction
similar to SSE2's _mm_movemask_epi8 (PMOVMSKB) which extracts one bit
from each byte and stores it in a register. We used that in the UTF-8
code to detect bytes with the highest bit set. To compensate, we used
two alternatives:

1) AND the comparison result with a vector containing {128, 64, ...,1 }
   Do 3 parallel-adds (VPADD.I8), which will make the mask propagate
   to the lowest component in the vector.
   Trick found in: http://hilbert-space.de/?p=22 (comment 16-17)

2) Extract the two words from the doubleword Neon register and do the
   work in ARM assembly.

It looks like the latter version is performing better.