qmake vcxproj generator: fix non flat projects for VS2010
The filters for files in the non flat projects will start with the
base filter name and then the path to the file. So the detection of
the filter needs to be done by the start of the filter namse so that
we correctly add the compile commands for the added files.

Task-number: QTBUG-13928
Reviewed-by: Joerg Bornemann