Fix bugs in tst_qnetworkreply
It used a unix way to delete local servers. Changed this to the cross
platform QLocalServer::removeServer API

It reported XPASS on symbian due to SRCDIR being a relative path not
absolute. (added a check for this, as SRCDIR could be relative on windows)

Also fixed a couple of compiler warnings

Reviewed-by: Martin Petersson