GraphicsViews: Focused widget is not visible after orientation change
If there is a graphicsview with input widget focused and splitview is
open, and we rotate the device from Portrait to Landscape
(or vice versa), the input widget is no longer visible in the
splitview rect. This only happens if there are scrollbars visible
in the view.

Reason is that the input widget is tried to ensure visibility for
before window surface is valid. In these cases, make the input widget
visibility re-location happen after window server indicates that the
view is now visible.

Task-number: QTBUG-17576
Reviewed-by: Guoqing Zhang