Call FixNativeOrientation on Symbian for certain fullscreen qml views.
For fullscreen QDeclarativeViews that lock their orientation to the
native orientation of the device an additional performance gain can
be achieved by calling FixNativeOrientation on the underlying RWindow.
Such apps will either work in a single orientation only, or they will
handle rotation themselves, by using QPainter transformations and
opting out from the standard Avkon auto-rotation. In any of these cases
there is no need for the lower level layers of the system to handle
graphics surface rotation, the app is "fixed" to the native orientation
from their point of view, therefore it is safe to call
FixNativeOrientation().

Task-number: QTBUG-17742
Reviewed-by: Jason Barron