Switched back to using symlinks for all mkspecs on Linux.
This is a partial revert of 7a19192fb418e40732ad72e0d4aaf83527958ded.

The original special case was there to accomodate the fact that
the Symbian master code line does not always build on a symlink
capable file system. However, if we don't use symlinks, qmake's code
to check for the symbian-sbsv2 scope fails if the mkspecs are not
symlinks, so if anyone wants to bring this change back, qmake must be
fixed first.

Strictly speaking it is not necessary to remove both cp's as is done
in this patch, but there is little point in keeping one and not the
other.

Task:     QTBUG-17736
RevBy:    Miikka Heikkinen