QTabWidget tabs are resized in wrong way when device rotates
This used to work so that, style would check from actual AVKON
component (navipane) what its size is. Unfortunately, Qt was later
changed so that screen furniture is only created on-demand and thus
asking navipane's size started to fail. Which lead to situation that
tab size would "grow" after navipane was created and tabbar
was re-drawn.

As a fix, correct the pixel metrics with correct values for nHD
and remove size query from native component. The other pixel metrics
for other layouts were correct.

Task-number: QTBUG-16984
Reviewed-by: Tomi Vihria