Moved the QML Observer Service and related functionality into Qt
This code was previously developed as part of Qt Creator in
share/qtcreator/qml/qmljsdebugger/

Moving it into Qt will allow us to simplify the setup required before
you can debug QML applications.

To avoid adding too much weight to the QtDeclarative module, a
declarativeobserver plugin was introduced that contains the
QDeclarativeViewObserver and related classes. The
QDeclarativeObserverService is just a stub service that loads this
plugin once a QML debugging client connects. The plugin implements the
QDeclarativeObserverInterface

A QJSDebugService was separated out of QJSDebuggerAgent, so
that the service can be active while the agent is instantiated lazily.
Each QDeclarativeEngine adds itself to the QJSDebugService. Currently
only the first one is used when instantiating the agent.

QDeclarativeObserverService is hooked into QDeclarativeView, with the
view registering itself to the service, allowing the
QDeclarativeViewObserver to be created for the view once somebody
connects to the service. Again, only the first view is used at the
moment.

Change-Id: Ib50579c6d24361c2b39528e5556410d3446c2e90
Reviewed-by: Martin Jones
Reviewed-by: Michael Brasser