Fix for tst_QEventLoop::processEventsExcludeSocket
Depending on the OS scheduling, the TCP data sent and flushed may not
have been looped back to trigger the socket's ready to read notification
before the test called processEvents.
This caused a failure on CI system windows 7.
Added a short delay before processEvents to allow the OS to catch up.

Reviewed-By: Markus Goetz