Don't pre-check if directory exists as it's unnecessary
On Posix platforms, when attempting to atomically create file and obtain
descriptor, we only continue if we get an EEXIST error. If the path
points to a non-existing directory, we should get ENOENT and will fail
as well.

On Windows, the check was already ignored. We would detect the file does
not exist, attempt to obtain a handle upon return to
QTemporaryFileEngine::open and fail there without running "for a *very*
long time".

Checking if the directory exists beforehand is not only unnecessary it
also constitutes a race condition. Trusting the result of this check is
dangerous and doubly useless.

Reviewed-by: Olivier Goffart
Reviewed-by: Robin Burchell