Fix the leak of QAdoptedThread and all dependent objects
if there is a QEventDispatcher in QAdoptedThread.

QAdoptedThread is supposed to be destroyed when the QThreadData is
destroyed. But QThreadData is ref-counted, and QEventDispatcher holds
a reference to it.  QEventDispatcher is supposed to be destroyed
in QThreadPrivate::finish, which is supposed to be called from the
destructor of QAdoptedThread.

There is a circular dependence.  We break it by calling finish in the
callback that is called when the thread is finished.

Task-number: QTBUG-17986
Reviewed-by: Brad