Symbian workaround for networkselftest
The frox ftp proxy closes connections using TCP RST instead of a normal
TCP FIN. On symbian, this causes the received but unacknowledged
preceding packet to be discarded (the "221 Goodbye" ftp response).
As QNAM and QFTP don't care about this, the network self test won't
either. (only for symbian, as windows and unix TCP/IP stacks retain
the unacknowledged packet in the receive queue)

Reviewed-by: Markus Goetz