QMeeGoLivePixmapData: Verify dimensions of locked pixmap
When a switch from HW rendering to SW rendering is done, all EGL
surfaces backing a live pixmap become invalid because EGL is
terminated. Thanks to a recent patch, all live images are now
automatically invalidated when SW rendering is activated.

Even with this mechanism in place, it is possible for the underlying EGL
surface to be different than what the live pixmap is expecting. This can
happen, for instance, if the X pixmap gets destroyed and a different X
drawable is created using the same XID. Even though this is an unlikely
scenario, it can result in a complete system failure if the surface
dimensions do not match and the live pixmap user ends up writing over
internal graphics driver structures.

This patch adds a safeguard which disallows locking EGL surfaces with
non-matching dimensions.

Fixes Harmattan bug NB#237138.

Merge-request: 2583
Reviewed-by: Samuel Rødal <samuel.rodal@nokia.com>