Make qhttpnetworkconnection autotest work with the new test server
Loosen requirements for generated error pages (these vary with exact
version and locale of the web server). Don't care what the length is,
but if a content-length header is sent then it should match the size
of the error page data. (apache2 does send content-length for errors)

Don't fail the gzip test if the server doesn't send a content-length
header. That's a server bug rather than Qt. (apache2 doesn't send)

Reviewed-by: Markus Goetz