Fix for failing autotest QToolBar/Symbian
QS60Style was calculating the minimum size of a toolbutton inside
a toolbar based on number of visible buttons. Of course, when the
non-visible buttons become visible and we are limiting the toolbar
width to screen width it fails (since the orginally visible button is
already claiming to take whole width of toolbar and now-visible button
is taking half the button width).

As a fix, calculate minimum size based on number buttons only,
irregardless of their visibility.

Task-number: QTBUG-17777
Reviewed-by: Miikka Heikkinen