Added preserved contents property to QWindowSurface.
Previously we used hasPartialUpdateSupport() to deduce whether the
window surface's contents are preserved when flush is called or not, but
that limits us from fully supporting platforms which might allow partial
updates even though the surface contents aren't preserved on flush.

If hasPreservedContents() returns false, that means a flush of a region
is never allowed without first painting to that region. If
hasPartialUpdateSupport() returns false, the whole window surface must
always be painted whenever an update is requested. Note that
hasPreservedContents() will typically imply hasPartialUpdateSupport(),
but not vice versa.

Reviewed-by: Bjørn Erik Nilsen