Make the QIconvCodec on Unix not enforce the Latin1 codec.
Originally, the QIconvCodec created and disposed of the iconv_t
structure at every turn. In Qt 4.5, we started saving it for later in
a thread-local storage.

We had to introduce a fix to make sure that we didn't create the
iconv_t structure before the setlocale(LC_ALL, "") call was made,
though: otherwise, we'd keep forever an iconv_t that pointed to the
wrong encoding.

So now simply restore the Qt 4.4 behaviour: create and dispose of the
iconv_t structure if we're called before the QCoreApplication
constructor is run or after the static destructors are run.

Note: this means QIconvCodec will probably default to US-ASCII when
run before QCoreApplication, not Latin 1. Non-ASCII characters
(anything with the high bit set) will fail to convert.

Task-number: QTBUG-15229?
Reviewed-by: Denis Dzyubenko