Use region name as a locale name on Mac.
This changes the behavior as before we were returning the name that we read
from System Preferences -> Language which represents the language for UI
translation.

Now that we have QLocale::uiLanguages() we don't need a hack that returned UI
language as a locale name.

This essentially reverts adb6eea3a135707a220c0b1c1cf8266321af9309.

Reviewed-by: con