look for makespec in the build dir first
it's more natural to look into the build dir before the source dir, and
it's what the qmake-generated makefiles mean when re-invoking qmake.

specifically, this works around the problem that relative paths with
excess ".."s pointing below the root are happily ignored and thus truly
bizarre makespec paths may be constructed by the qmake re-invocations if
the source dir is less nested than the build dir.

Task-number: QTBUG-9817
Reviewed-by: mariusSO