QTemporaryFile: Locate placeholder before re-encoding
Previously, the placeholder was identified in unicode-based QString, and
the offset was assumed to be correct after conversion to the local 8-bit
encoding. This assumption is obviously wrong for variable-sized
encoding, such as utf-8.

With this change, the placeholder is still identified in the original
QString, but conversion to 8-bit encoding is done in a step-wise fashion
and offsets recomputed. This allows us to drop the intermediary argument
verification in _gettemp.

Task-number: QTBUG-4796
Reviewed-by: Robin Burchell