Fix assert fail when debug log is enabled
Enough debug logging allows time for the proxy server to close the http
connection after sending a 407 response. This errors the connection,
but then the queued _q_startNextRequest is immediately run, causing an
assertion failure.
Changed this to a soft failure which allows the error to propagate
rather than crashing.
Test case is tst_qnetworkreply::ioPostToHttpFromSocket

Reviewed-by: Markus Goetz