Update address reporting code
When an IP6 address contains a valid IP4 address, report the address
in IP4 format (this is needed for SOCKS5 socket engine)
When binding Any or AnyIPv6, create a dual mode socket (KAfUnspec),
but report the local address as being the same as what the user
requested. (by default, symbian returns ::0 for the dual mode socket,
which causes problems for SOCKS and UDP code)

The intent is that most applications written for IP4 can work
transparently in an IP6 environment.
QTcpServer or QUdpSocket can accept either IP4 or IP6 from the same
socket.

Reviewed-by: Markus Goetz