opengl2: Make maximum cached glyph size configurable
As the DPI of displays continues to rise, especially on mobile devices,
the maximum glyph size of 64 pixels for the glyph cache has become a
limitation. This problem is made worse by the fact that when the maximum
glyph size is exceeded, the OpenGL paint engine falls back to using
rasterized geometry for glyph rendering. This does not produce
acceptable quality if the OpenGL rendering surface lacks support for
multisampling.

This patch offers a solution to the problem by making the
cached glyph size configurable at build time. This way the limit can be
set according to the capabilities of the target hardware.

Signed-off-by: Sami Kyöstilä <sami.kyostila@nokia.com>

Merge-request: 1131
Reviewed-by: Samuel Rødal <samuel.rodal@nokia.com>