Cocoa: p1 bugfix, add widget flag MacNoCocoaChildWindow
This problem has been known for a long time, but a good solution has
never been found. The problem is that a child window should always
stay on top of it's parent, if not for anything else than secure that
a modal child does not block input while hiding behind the parent at
the same time. The only sensible solution found to ensure this in the
Cocoa port is to use Cocoa child windows. But this API has a sad side
effect; it will move the child along with the parent when the parent
is moved on screen. This is something it seems we have to live with.
But for those users that wants to handle this issue otherwise, we now
add a widget flag to switch this off.

Task-number: QTBUG-11481
Reviewed-by: msorvig