QApplication does not define flag for "single touch"
Qt application was previously not setting a native framework's flag to
indicate that is supports "single touch" (i.e. activating an item would
only need one tap, instead of one tap to highlight and another to
activate) for Sym^3.

This was causing the native dialogs launched from within the
application to behave as Sym^1 dialogs.

As a fix, set the native flag when running Qt in Sym^3.

Task-number: QT-4569
Reviewed-by: Jani Hautakangas