Documentation Amelioration
A minor typo in a documentation snippet left the 'c' off of 'can',
leading to the modal verb being misparsed as a singular pronoun and
resulting in inefficient communication of the intent. The wayward letter
has since been returned to its accustomed place.

Complete and thorough description contained in the commit message was
added to appease the git hook that didn't like 'Doc Fix'.

Change-Id: I72d799f15f8fed4c470443bdbbd4344ded3dbd80