Fixed deployment problems with MSVC 2005 and 2008 SP1
Notably, DLLs bundled with MSVC Express SP1 releases are only the ones
from the SP1, so it is a good idea to force binding to the latest
release of the runtime to avoid deployment problems.

Keep in mind that it is easier to bundle a private runtime for the
software in an setup EXE rather than lauching a runtime setup from the
software setup.

Task-number: QTBUG-8215
Merge-request: 1171
Reviewed-by: Joerg Bornemann <joerg.bornemann@nokia.com>