Mac: p2 combobox regression fix
The combobox popup on mac fails to re-highlight the last item in the
list when then mouse leaves the popup, and re-enter. We tried to fix
this in 780b4d84, but is turned out to break the cleanlooks style.
For this second try, we allow the code to set the index under mouse
when the mouse hovers an item in the list, even if the index is
current. It seems to cause low overhead, and should be safe.

Task-number: QTBUG-17738
Reviewed-by: jbache