Improve Flickable dynamics and allow platform specific tweaking.
Flickable's previous overshoot behavior was to continue past the
boundary without any additional friction (deceleration). It now
decelerates faster when overshooting.

On touch screens in particular, the last points when a touch point
is released are unreliable, resulting in sporadic flick velocities.
It also now allows the number of samples used for velocity calclations
to be tuned.

It is now easy to tune the Flickable behavior for different platforms.

Change-Id: I24142a50be1fde2f8877e359e30b8efcdd1f7d5c
Task-number: QTBUG-10894, QTBUG-16388, QTBUG-17830
Reviewed-by: Michael Brasser