Not possible to enter negative values to widgets with numeric fields
Currently it is not possible to enter negative number to QSpinBox (or
to any other QWidget which is handling numeric input only) in touch
devices. This is due to that AVKON virtual keyboards have
negative sign (-) in special character table(s) and it is not
accessible by default from QWidgets, unless you define widget's input
hints as ImhFormattedNumbersOnly AND ImhDialableCharactersOnly
(or non-numeric).

With this fix, default special character table is added for widgets
with numeric input mode AND either ImhFormattedNumbersOnly or
ImhDialableCharactersOnly.

By default user is still unable to enter nagative values to QSpinBox,
but developer can set inputMethodHint to work around the issue:

    widget->setInputMethodHints(Qt::ImhFormattedNumbersOnly);

Task-number: QTBUG-17461
Reviewed-by: Miikka Heikkinen