Have the backing store destroyed also in special Symbian scenarios.
The backing store tracker's registerWidget and unregisterWidget
functions are called when EPartiallyVisible and ENotVisible events
come from WSERV. However if an application sends its window group
to background right after caling show() and before entering the
event loop, there is a chance that all the application will receive
is an ENotVisible event, leading to calling unregisterWidget() without
a previous registerWidget(). In this case the backing store was
not destroyed and the application was consuming GPU memory even while
it was staying in background. This patch makes unregisterWidget()
not to check the widget's presence in the m_widgets set, instead
the condition for deleting the backing store is solely an empty set.

Task-number: QTBUG-18493
Reviewed-by: Gareth Stockwell