Cocoa: p1 bug fix: fix auto test regressions
Ref: 32228c4f2b3419a35d1623377050ef72edf73c92
It seems that the change above broke some auto tests, which
revealed a true problem. When it comes to modal dialog, children
still needs to be stacked on top of modal parents, as they the
user cannot use the mouse to raise it. So rather than removing
subWindowStacking fully, we narrow it even further down to
only be used for children of modal dialogs. All in all, this
is close to removing it, but still us it for certain corner cases.

Task-number: QTBUG-11481
Reviewed-by: msorvig