Make sure removed QTextBlock is invalid
If the block is removed from document block map, we will mark the
right node to the current head->freelist index, but it shouldn't
be accessed directly, otherwise it can cause crash because of
uninitialized node.

Hence we need to check if a node index is equal to current freelist
index. If so, it cannot be a valid block.

Task-number: QTBUG-18500
Reviewed-by: Eskil