tst_QStyle::drawItemPixmap test case fails on Symbian^3
The autotest assumes that created QPixmap are exactly same irregardless
how the pixmaps were created. However, there is no guarantee that
pixmaps that look the same (i.e. in this case green rectangles)
are "same". QPixmap is platform dependent and might have e.g.
optimized format in some of the platforms.

Task-number: QT-4805
Reviewed-by: Jani Hautakangas