Two QListView autotests do not pass on Symbian^3 releases
tst_QListView::taskQTBUG_2678_spacingAndWrappedText() fails because:
QS60Style adds to the itemview item content size margins and empty
space, which shouldn't be part of the content size.
As a fix, remove these.

taskQTBUG_435_deselectOnViewportClick() fails because:
Sending a click to a selected itemview item, when it should be sent
to below that specific one. It was using hardcoded (center + 20) and
autotest assumed that this would be outside of first item.
In S60 with touch support, the itemview items are rather tall
(49 pixels). As a fix, autotest now uses calculated value, which
ensures that click is sent to outside first item.

Task-number: QT-4810
Reviewed-by: Tomi Vihria