Use FixNativeOrientation on Symbian when the application wants it.
Applications can request the usage of RWindow::FixNativeOrientation()
by setting the new Qt::WA_SymbianNoSystemRotation attribute for their
fullscreen top-level widget (which in practice will either be a
graphics or declarative view). This will fix the underlying EGL
window surface and the QWidget dimensions to the dimension of the
native orientation of the device (typically portrait). The default
auto-rotation can be left enabled, however it will be up to the
application to rotate the drawing. Global notifications, VKB, etc.
will still appear in the proper orientation. Another benefit is
improved performance in the non-native orientation.

Task-number: QTBUG-17742
Reviewed-by: Jason Barron
Reviewed-by: Jani Hautakangas