Have the backing store destroyed also in special Symbian scenarios.
The backing store tracker's registerWidget and unregisterWidget
functions are called when EPartiallyVisible and ENotVisible events
come from WSERV. However if an application sends its window group
to background right after caling show() and before entering the
event loop, there is a chance that all the application will receive
is an ENotVisible event, leading to calling unregisterWidget() without
a previous registerWidget(). In this case the backing store was
not destroyed and the application was consuming GPU memory even while
it was staying in background. This patch ensures registerWidget() is
called always before unregisterWidget() in the Symbian-specific
event handler for ENotVisible, and replaces the previous patch,
a12d41076919a133e63de63dff5c1a131a0564e4, which caused regression
in autotests.

Task-number: QTBUG-18493
Reviewed-by: Gareth Stockwell