QDeclarativeEnginePrivate: fix and encapsulate the usage of scopeChainValue
Reduce the places that we make a direct reference to the
scopeChainValue (and the "-3" position) by providing a higher level
function that gets the scopeNode of the Declarative Context that
is in the current execution.

This commit also fixes the offset of the wanted nodes in the scope
chain. In the long term we still need to get rid of this scope chain
direct manipulation if possible.

Reviewed-by: Olivier Goffart