Splitview: opening a context menu should not re-position of focusItem
When splitview is open and user long taps an text input item,
a context menu is opened. This leads to some drawing operations on
splitview implementation, even though these should be skipped.
The focusItem might even be re-positioned into incorrect position
due to the failure.

As a fix, check if the focusItem is already positioned correctly
and avoid doing anything when it already is.

Task-number: QTBUG-18738
Reviewed-by: Guoqing Zhang