Fix qstringbuilder4 test on mac
The behaviour is different with or without QT_FAST_OPERATOR_PLUS
And since on old gcc it is not defined, then the test fails.

With the normal operator+, the bytearray will first be converted to
a QString, and loose everything that is after \0, but now, with
QStringBuilder, we do not do QString conversion anymore, so no data
is lost.

We do not want to specify the behaviour here, so we remove the test.

Reviewed-by: Joao