Splitview flag should also be updated to inputcontext
Currently splitview support is defined in exported global private method.
There is no guarantee that input context is created, or updated when
the API is called. This is problematic, since the client need to find
a correct place and time when to call the API.

As a change, update existing input context when the API is called.
Also, take into account the splitview flag when updating input context.
This ensures that client can enforce the flag use correctly.

Task-number: QTBUG-18715
Reviewed-by: Guoqing Zhang