Resolve unqualified attached properties correctly
When resolving unqualified attached properties, we should use the scope
object, not the context object.  Otherwise they will always resolve to the
root object of the context, regardless of where they are written.

In this example,

QtObject {
    id: root
    QtObject {
        id: me
        property int a: AttachedObject.x
    }
}

the attached object should be loaded on the "me" object, not the "root"
object.

Change-Id: I386f886f62df7b8020c3ff703cdfc891d5739713
Reviewed-by: Martin Jones